pyproject.toml
[tool.poetry]
name = "rnmp-homework3"
version = "0.1.0"
description = ""
authors = ["lupusruber <petarpro234@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
kafka-python = "^2.0.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
scikit-learn = "^1.6.0"
xgboost = "^2.1.3"
black = "^24.10.0"
pyspark = "^3.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
